+++ HISTORY.md
@@ -2,6 +2,7 @@
 ==================
 
 * Completely rewritten module core in ES6 while retaining compatiblity with last Node.js LTS (which is `4.5.0` at the moment of writing)
+* Received redis messages can now be manipulated before pushing them to the stream
 * Switched to `ioredis` as Redis client module
 * Switched to `AVA` for tests
 * Using SEMVER